requirements.txt
numpy=2.2.1
pillow=11.1.0
prettytable=3.12.0